hand_test2

+++ 2015-08-16/src/main/java/com/hand/Bean/IdEntity.java
@@ -2,13 +2,13 @@ package com.hand.Bean;
 
 public abstract class IdEntity {
 	
-	protected Long id;
+	protected int id;
 
-	public Long getId() {
+	public int getId() {
 		return id;
 	}
 
-	public void setId(Long id) {
+	public void setId(int id) {
 		this.id = id;
 	}
 }
